Add finalizer and atexit callbacks to worker
The finalizer cannot do a task switch, so it's necessary to create a new
task inside it. That task might not actually be executed when Julia
exits, so a redundant `atexit` callback is also necessary.

Update the `stop` function to check whether the process is still
running. The pattern `isrunning(w) && stop(w)` was becoming very common,
so this is clearly needed. If the user wants to stop a process, and the
process is already dead, why should there be an exception? The end
result is what the user wanted: no active process.